refactor(settings): extract archived branch grouping
